% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pl_do_gist}
\alias{pl_do_gist}
\title{Post a file as a Github gist}
\usage{
pl_do_gist(gist, description = "", public = TRUE, browse = TRUE)
}
\arguments{
\item{gist}{An object}

\item{description}{brief description of gist (optional)}

\item{public}{whether gist is public (default: TRUE)}

\item{browse}{If TRUE (default) the map opens in your default browser}
}
\description{
Post a file as a Github gist
}
\keyword{internal}

