% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_places.R
\name{pl_places}
\alias{pl_places}
\title{Get data for a place given a place ID}
\usage{
pl_places(place_id, ...)
}
\arguments{
\item{place_id}{(integer/numeric) A place ID}

\item{...}{Curl options, see \code{\link[curl]{curl_options}}}
}
\description{
Get data for a place given a place ID
}
\examples{
\dontrun{
pl_places(place_id = 462471)
}
}

