% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pl_search}
\alias{pl_search}
\title{Search for a place, name or X.}
\usage{
pl_search(query, ...)
}
\arguments{
\item{query}{A place ID}

\item{...}{Further args passed on to \code{agrep}}
}
\description{
This function searches a locally cached file of their data catalog.
You can update the cache if you want. See examples.
}
\examples{
\dontrun{
pl_search(query='Rome')
}
}

