% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pl_search_loc}
\alias{pl_search_loc}
\alias{pl_search_names}
\alias{pl_search_places}
\title{Search for a place, name or location.}
\usage{
pl_search_loc(query = NULL, path = "~/.pleiades/", ...)

pl_search_names(query, path = "~/.pleiades/", ...)

pl_search_places(query, path = "~/.pleiades/", ...)
}
\arguments{
\item{query}{A place ID}

\item{path}{(character) Path to cache data in.}

\item{...}{Further args passed on to \code{\link[dplyr]{tbl}}}
}
\description{
This function searches a locally cached file of their data catalog. See examples.
}
\examples{
\dontrun{
pl_search_loc("SELECT * FROM locations limit 5")
pl_search_names(query = "SELECT * FROM names limit 5")
pl_search_places("SELECT * FROM places limit 5")
}
}

