% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pl_status}
\alias{pl_status}
\title{Get Pleiades status data, number of places, number of locations, number of names.}
\usage{
pl_status(...)
}
\arguments{
\item{...}{Curl debugging arguments, see \code{httr::config}}
}
\description{
Get Pleiades status data, number of places, number of locations, number of names.
}
\examples{
\dontrun{
pl_status()
}
}

