% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleiades-package.r
\docType{package}
\name{pleiades}
\alias{crul}
\alias{pleiades}
\alias{pleiades-package}
\title{\strong{pleiades}}
\description{
\strong{pleiades}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}

