% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pl_cache}
\alias{pl_cache}
\alias{pl_cache_clear}
\title{Cache data locally for later usage.}
\usage{
pl_cache(path = "~/.pleiades/", overwrite = TRUE, force = FALSE, ...)

pl_cache_clear(path = "~/.pleiades/", which = NULL, prompt = TRUE)
}
\arguments{
\item{path}{(character) Path to cache data in.}

\item{overwrite}{(logical) Overwrite existing files?}

\item{...}{Further args passed on to \code{\link[httr]{GET}}}

\item{which}{(character) One of locations, names, or places.}

\item{prompt}{(logical) Prompt before clearing all files in cache? No prompt used when DOIs
assed in. Default: TRUE}
}
\description{
Cache data locally for later usage.
}
\examples{
\dontrun{
pl_cache()
pl_cache(force = TRUE)

# clear all files
pl_cache_clear()

# clear a single file
pl_cache_clear(which = "locations")
pl_cache_clear(which = "places")
pl_cache_clear(which = "names")
}
}

