% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pl_gist}
\alias{pl_gist}
\title{Make an interactive map to view in the browser as a Github gist}
\usage{
pl_gist(x, outfile = NULL, description = "", browse = TRUE)
}
\arguments{
\item{x}{Output from pl_places}

\item{outfile}{File name (without file extension) for your geojson file. Default is 'gistmap'.}

\item{description}{Description for the Github gist, or leave to default (=no description)}

\item{browse}{If TRUE (default) the map opens in your default browser.}
}
\value{
Creates a gist on your Github account, and prints out where the geojson file was
written on your machinee, the url for the gist, and an embed script in the console.
}
\description{
You will be asked ot enter you Github credentials (username, password) during
each session, but only once for each session. Alternatively, you could enter
your credentials into your .Rprofile file with the entries

\itemize{
 \item options(github.username = 'your_github_username')
 \item options(github.password = 'your_github_password')
}

then \code{pl_gist} will simply read those options.

\code{pl_gist} has modified code from the rCharts package by Ramnath Vaidyanathan
}
\examples{
\dontrun{
xxx <- pl_place(place_id=579885)
pl_gist(x=xxx)
}
}

