% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pl_places}
\alias{pl_places}
\title{Get data for a place given a place ID}
\usage{
pl_places(place_id, ...)
}
\arguments{
\item{place_id}{A place ID}

\item{...}{Curl debugging arguments, see \code{httr::config}}
}
\description{
Get data for a place given a place ID
}
\examples{
\dontrun{
pl_places(place_id=579885)
}
}

