% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pl_search_loc}
\alias{pl_search_loc}
\alias{pl_search_names}
\alias{pl_search_places}
\title{Search for a place, name or location.}
\usage{
pl_search_loc(query = NULL, path = "~/.pleiades/", ...)

pl_search_names(query, path = "~/.pleiades/", ...)

pl_search_places(query, path = "~/.pleiades/", ...)
}
\arguments{
\item{query}{A place ID. If left NULL, returns the table, which is of class \code{tbl},
which can then be passed on to other \code{dplyr} functions.}

\item{path}{(character) Path to cache data in.}

\item{...}{Further args passed on to \code{\link[dplyr]{tbl}}}
}
\description{
This function searches a locally cached file of their data catalog. See examples.
}
\details{
On the first query if not run before, the function takes a bit to load the raw csv
data, then create a SQLite database, and create the pointer to it. Subsequent calls should
be very fast.
}
\examples{
\dontrun{
pl_search_loc("SELECT * FROM locations limit 5")
pl_search_names(query = "SELECT * FROM names limit 5")
pl_search_places("SELECT * FROM places limit 5")
}
}

