% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_status.R
\name{pl_status}
\alias{pl_status}
\title{Get Pleiades status data, number of places, number of locations,
number of names}
\usage{
pl_status(...)
}
\arguments{
\item{...}{Curl options, see \code{curl::curl_options()}}
}
\description{
Get Pleiades status data, number of places, number of locations,
number of names
}
\examples{
\dontrun{
pl_status()
}
}
